[ADD] payroll_salary_rule_python, removed from payroll
[FIX] hr_salary_rule_python, moved from payroll (tests)

[FIX] payroll_salary_rule_python (demo, readme, pre-commit)

[FIX] payroll, payroll_account, payroll_salary_rule_python

[FIX] payroll/tests/common

[FIX] payroll/tests

[FIX] payroll/tests/test_payslip_flow
