+++ payroll/demo/hr_payroll_demo.xml
@@ -28,48 +28,21 @@
         <field name="name">Company Contribution</field>
         <field name="code">COMP</field>
     </record>
+    <!-- Salary Rules -->
     <record id="hr_rule_basic" model="hr.salary.rule">
         <field name="name">Basic Salary</field>
         <field name="sequence" eval="1" />
         <field name="code">BASIC</field>
         <field name="category_id" ref="payroll.BASIC" />
         <field name="condition_select">none</field>
-        <field name="amount_select">code</field>
-        <field name="amount_python_compute">
-result = contract.wage
-        </field>
-    </record>
-    <record id="hr_rule_taxable" model="hr.salary.rule">
-        <field name="name">Gross</field>
-        <field name="sequence" eval="100" />
-        <field name="code">GROSS</field>
-        <field name="category_id" ref="payroll.GROSS" />
-        <field name="condition_select">none</field>
-        <field name="amount_select">code</field>
-        <field name="amount_python_compute">
-result = categories.BASIC + categories.ALW
-        </field>
-    </record>
-    <record id="hr_rule_net" model="hr.salary.rule">
-        <field name="name">Net Salary</field>
-        <field name="sequence" eval="200" />
-        <field name="code">NET</field>
-        <field name="category_id" ref="payroll.NET" />
-        <field name="condition_select">none</field>
-        <field name="amount_select">code</field>
-        <field name="amount_python_compute">
-result = categories.BASIC + categories.ALW + categories.DED
-        </field>
-        <field name="register_id" ref="contrib_register_employees" />
+        <field name="amount_select">fix</field>
+        <field name="amount_fix" eval="1000.0" />
     </record>
     <!-- Salary Structure -->
     <record id="structure_base" model="hr.payroll.structure">
         <field name="code">BASE</field>
         <field name="name">Base for new structures</field>
-        <field
-            eval="[(6, 0, [ref('hr_rule_basic'), ref('hr_rule_taxable'),ref('hr_rule_net')])]"
-            name="rule_ids"
-        />
+        <field name="rule_ids" eval="[(6, 0, [ref('hr_rule_basic')])]" />
         <field name="company_id" ref="base.main_company" />
     </record>
     <!-- Contribution Register -->
@@ -143,35 +116,13 @@ result = categories.BASIC + categories.ALW + categories.DED
         <field name="name">Meal Voucher</field>
         <field name="sequence" eval="16" />
     </record>
-    <record id="hr_salary_rule_sales_commission" model="hr.salary.rule">
-        <field name="amount_select">code</field>
-        <field name="code">SALE</field>
-        <field name="category_id" ref="payroll.ALW" />
-        <field name="name">Get 1% of sales</field>
-        <field name="sequence" eval="17" />
-        <field name="amount_python_compute">
-result = ((inputs.SALEURO and inputs.SALEURO.amount) + (inputs.SALASIA and
-inputs.SALASIA.amount)) * 0.01
-        </field>
-    </record>
-    <!-- Rule Inputs -->
-    <record id="hr_rule_input_sale_a" model="hr.rule.input">
-        <field name="code">SALEURO</field>
-        <field name="name">Sales to Europe</field>
-        <field name="input_id" ref="hr_salary_rule_sales_commission" />
-    </record>
-    <record id="hr_rule_input_sale_b" model="hr.rule.input">
-        <field name="code">SALASIA</field>
-        <field name="name">Sales to Asia</field>
-        <field name="input_id" ref="hr_salary_rule_sales_commission" />
-    </record>
     <!-- Salary Structure -->
     <record id="structure_001" model="hr.payroll.structure">
         <field name="code">ME</field>
         <field name="name">Marketing Executive</field>
         <field
-            eval="[(6, 0, [ref('hr_salary_rule_houserentallowance1'),
-ref('hr_salary_rule_convanceallowance1'),ref('hr_salary_rule_professionaltax1'),ref('hr_salary_rule_providentfund1')])]"
+            eval="[(6, 0, [ref('payroll.hr_salary_rule_houserentallowance1'),
+ref('payroll.hr_salary_rule_convanceallowance1'),ref('payroll.hr_salary_rule_professionaltax1'),ref('payroll.hr_salary_rule_providentfund1')])]"
             name="rule_ids"
         />
         <field name="company_id" ref="base.main_company" />
@@ -181,7 +132,7 @@ ref('hr_salary_rule_convanceallowance1'),ref('hr_salary_rule_professionaltax1'),
         <field name="code">MEMD</field>
         <field name="name">Marketing Executive for Marc Demo</field>
         <field
-            eval="[(6, 0, [ref('hr_salary_rule_ca_demo'), ref('hr_salary_rule_meal_voucher')])]"
+            eval="[(6, 0, [ref('payroll.hr_salary_rule_ca_demo'), ref('payroll.hr_salary_rule_meal_voucher')])]"
             name="rule_ids"
         />
         <field name="company_id" ref="base.main_company" />
@@ -205,8 +156,8 @@ ref('hr_salary_rule_convanceallowance1'),ref('hr_salary_rule_professionaltax1'),
         <field name="name">Marketing Executive Contract</field>
         <field name="date_start" eval="time.strftime('%Y-%m')+'-1'" />
         <field name="date_end" eval="time.strftime('%Y')+'-12-31'" />
-        <field name="struct_id" ref="payroll.structure_001" />
-        <field name="employee_id" ref="hr_employee_payroll" />
+        <field name="struct_id" ref="structure_001" />
+        <field name="employee_id" ref="payroll.hr_employee_payroll" />
         <field name="notes">Default contract for marketing executives</field>
         <field eval="4000.0" name="wage" />
     </record>
@@ -214,9 +165,10 @@ ref('hr_salary_rule_convanceallowance1'),ref('hr_salary_rule_professionaltax1'),
         <field name="name">Contract For Marc Demo</field>
         <field name="date_start" eval="time.strftime('%Y-%m')+'-1'" />
         <field name="date_end" eval="time.strftime('%Y')+'-12-31'" />
-        <field name="struct_id" ref="payroll.structure_002" />
+        <field name="struct_id" ref="structure_002" />
         <field name="employee_id" ref="hr.employee_qdp" />
         <field name="notes">This is Marc Demo's contract</field>
         <field eval="5000.0" name="wage" />
     </record>
+
 </odoo>
