+++ payroll/models/hr_payslip.py
@@ -580,13 +580,11 @@ class HrPayslip(models.Model):
             "appears_on_payslip": rule.appears_on_payslip,
             "parent_rule_id": rule.parent_rule_id.id,
             "condition_select": rule.condition_select,
-            "condition_python": rule.condition_python,
             "condition_range": rule.condition_range,
             "condition_range_min": rule.condition_range_min,
             "condition_range_max": rule.condition_range_max,
             "amount_select": rule.amount_select,
             "amount_fix": rule.amount_fix,
-            "amount_python_compute": rule.amount_python_compute,
             "amount_percentage": rule.amount_percentage,
             "amount_percentage_base": rule.amount_percentage_base,
             "register_id": rule.register_id.id,
