+++ payroll/tests/test_payslip_flow.py
@@ -8,20 +8,6 @@ from .common import TestPayslipBase
 
 
 class TestPayslipFlow(TestPayslipBase):
-    def setUp(self):
-        super().setUp()
-
-        self.test_rule = self.SalaryRule.create(
-            {
-                "name": "Test rule",
-                "code": "TEST",
-                "category_id": self.categ_alw.id,
-                "sequence": 5,
-                "amount_select": "code",
-                "amount_python_compute": "result = contract.wage",
-            }
-        )
-
     def test_00_payslip_flow(self):
         """Testing payslip flow and report printing"""
 
@@ -56,15 +42,15 @@ class TestPayslipFlow(TestPayslipBase):
 
         # Check child rules shown in table by default
         child_line = richard_payslip.dynamic_filtered_payslip_lines.filtered(
-            lambda l: l.code == "NET_CHILD"
+            lambda l: l.code == "BASIC_CHILD"
         )
         self.assertEqual(
-            len(child_line), 1, "Child line found when flag desactivated (default)"
+            len(child_line), 1, "Child line found when flag deactivated (default)"
         )
 
         # Check parent line id value is correct
         parent_line = richard_payslip.dynamic_filtered_payslip_lines.filtered(
-            lambda l: l.code == "NET"
+            lambda l: l.code == "BASIC"
         )
         self.assertEqual(
             child_line.parent_line_id.code,
@@ -82,23 +68,10 @@ class TestPayslipFlow(TestPayslipBase):
 
         # Check child rules not shown in table after flag changed
         child_line = richard_payslip.dynamic_filtered_payslip_lines.filtered(
-            lambda l: l.code == "NET_CHILD"
+            lambda l: l.code == "BASIC_CHILD"
         )
         self.assertEqual(
             len(child_line), 0, "The child line is not found when flag activated"
-        )
-
-        # Find the 'NET' payslip line and check that it adds up
-        # salary + HRA + MA + SALE - PT
-        work100 = richard_payslip.worked_days_line_ids.filtered(
-            lambda x: x.code == "WORK100"
-        )
-        line = richard_payslip.line_ids.filtered(lambda l: l.code == "NET")
-        self.assertEqual(len(line), 1, "I found the 'NET' line")
-        self.assertEqual(
-            line[0].amount,
-            5000.0 + (0.4 * 5000) + (work100.number_of_days * 10) + 0.05 - 200.0,
-            "The 'NET' amount equals salary plus allowances - deductions",
         )
 
         # Then I click on the 'Confirm' button on payslip
@@ -162,28 +135,6 @@ class TestPayslipFlow(TestPayslipBase):
             "action_payslip_lines_contribution_register",
             context=context,
             our_module="payroll",
-        )
-
-    def test_contract_qty(self):
-
-        # I set the test rule to detect contract count
-        self.test_rule.amount_python_compute = (
-            "result = payroll.contracts and payroll.contracts.count or -1.0"
-        )
-        self.developer_pay_structure.write({"rule_ids": [(4, self.test_rule.id)]})
-
-        # I put all eligible contracts (including Richard's) in an "open" state
-        self.apply_contract_cron()
-
-        # I create an employee Payslip and process it
-        richard_payslip = self.Payslip.create({"employee_id": self.richard_emp.id})
-        richard_payslip.onchange_employee()
-        richard_payslip.compute_sheet()
-
-        line = richard_payslip.line_ids.filtered(lambda l: l.code == "TEST")
-        self.assertEqual(len(line), 1, "I found the Test line")
-        self.assertEqual(
-            line[0].amount, 1.0, "The calculated dictionary value 'contracts.qty' is 1"
         )
 
     def test_compute_multiple_payslips(self):
@@ -279,19 +230,13 @@ class TestPayslipFlow(TestPayslipBase):
         developer_rules |= self.rule_basic
         developer_rules |= self.rule_hra
         developer_rules |= self.rule_meal
-        developer_rules |= self.rule_commission
-        developer_rules |= self.rule_gross
         developer_rules |= self.rule_proftax
         developer_rules |= self.rule_child
-        developer_rules |= self.rule_net
         return developer_rules
 
     def _get_sales_rules(self):
         sales_rules = self.SalaryRule
         sales_rules |= self.rule_basic
-        sales_rules |= self.rule_commission
-        sales_rules |= self.rule_gross
-        sales_rules |= self.rule_net
         sales_rules |= self.rule_child
         return sales_rules
 
@@ -414,8 +359,9 @@ class TestPayslipFlow(TestPayslipBase):
                 "code": "ST",
                 "company_id": self.ref("base.main_company"),
                 "rule_ids": [
-                    (4, self.rule_commission.id),
                     (4, self.rule_basic.id),
+                    (4, self.rule_child.id),
+                    (4, self.rule_meal.id),
                 ],
             }
         )
@@ -432,7 +378,7 @@ class TestPayslipFlow(TestPayslipBase):
         payslip.onchange_struct_id()
         self.assertEqual(
             payslip._get_salary_rules(),
-            self.rule_commission | self.rule_basic,
+            self.rule_basic | self.rule_child | self.rule_meal,
             "Salary rules have changed to new structure",
         )
 
@@ -451,8 +397,9 @@ class TestPayslipFlow(TestPayslipBase):
                 "code": "ST",
                 "company_id": self.ref("base.main_company"),
                 "rule_ids": [
-                    (4, self.rule_commission.id),
                     (4, self.rule_basic.id),
+                    (4, self.rule_child.id),
+                    (4, self.rule_meal.id),
                 ],
             }
         )
@@ -479,12 +426,12 @@ class TestPayslipFlow(TestPayslipBase):
         payslips.onchange_struct_id()
         self.assertEqual(
             payslips[0]._get_salary_rules(),
-            self.rule_commission | self.rule_basic,
+            self.rule_basic | self.rule_child | self.rule_meal,
             "For 1st slip salary rules have changed to new structure",
         )
         self.assertEqual(
             payslips[1]._get_salary_rules(),
-            self.rule_commission | self.rule_basic,
+            self.rule_basic | self.rule_child | self.rule_meal,
             "For 2nd slip salary rules have changed to new structure",
         )
