+++ payroll/views/hr_salary_rule_views.xml
@@ -121,15 +121,6 @@
                                         nolabel="1"
                                     />
                                     <field
-                                        name="condition_python"
-                                        attrs="{'invisible':[('condition_select','!=','python')], 'required': [('condition_select','=','python')]}"
-                                        colspan="4"
-                                        widget="ace"
-                                        options="{'mode': 'python'}"
-                                        id="condition_python"
-                                        nolabel="1"
-                                    />
-                                    <field
                                         name="condition_range"
                                         attrs="{'invisible':[('condition_select','!=','range')], 'required':[('condition_select','=','range')]}"
                                     />
@@ -174,15 +165,6 @@
                                     <field
                                         name="amount_fix"
                                         attrs="{'invisible':[('amount_select','!=','fix')], 'required':[('amount_select','=','fix')]}"
-                                    />
-                                    <field
-                                        colspan="4"
-                                        name="amount_python_compute"
-                                        attrs="{'invisible':[('amount_select','!=','code')], 'required':[('amount_select','=','code')]}"
-                                        widget="ace"
-                                        options="{'mode': 'python'}"
-                                        id="amount_python_compute"
-                                        nolabel="1"
                                     />
                                     <field
                                         name="amount_percentage"
