+++ payroll_salary_rule_python/__manifest__.py
@@ -1,0 +1,21 @@
+# Copyright (C) 2022 Henrik Norlin (<henrik@appstogrow.co>).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+{
+    "name": "Salary Rule Python",
+    "version": "14.0.1.0.0",
+    "category": "Human Resources",
+    "website": "https://github.com/OCA/payroll",
+    "summary": "",
+    "license": "LGPL-3",
+    "author": "Henrik Norlin, Odoo Community Association (OCA)",
+    "depends": ["payroll"],
+    "data": [
+        "views/hr_salary_rule_views.xml",
+    ],
+    "demo": [
+        "demo/hr_payroll_demo.xml",
+    ],
+    "installable": True,
+    "maintainers": ["appstogrow"],
+}
