+++ payroll_salary_rule_python/demo/hr_payroll_demo.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <!-- Salary Rules -->
+    <record id="hr_rule_taxable" model="hr.salary.rule">
+        <field name="name">Gross</field>
+        <field name="sequence" eval="100" />
+        <field name="code">GROSS</field>
+        <field name="category_id" ref="payroll.GROSS" />
+        <field name="condition_select">none</field>
+        <field name="amount_select">code</field>
+        <field name="amount_python_compute">
+result = categories.BASIC + categories.ALW
+        </field>
+    </record>
+    <record id="hr_rule_net" model="hr.salary.rule">
+        <field name="name">Net Salary</field>
+        <field name="sequence" eval="200" />
+        <field name="code">NET</field>
+        <field name="category_id" ref="payroll.NET" />
+        <field name="condition_select">none</field>
+        <field name="amount_select">code</field>
+        <field name="amount_python_compute">
+result = categories.BASIC + categories.ALW + categories.DED
+        </field>
+        <field name="register_id" ref="payroll.contrib_register_employees" />
+    </record>
+    <record id="hr_salary_rule_sales_commission" model="hr.salary.rule">
+        <field name="amount_select">code</field>
+        <field name="code">SALE</field>
+        <field name="category_id" ref="payroll.ALW" />
+        <field name="name">Get 1% of sales</field>
+        <field name="sequence" eval="17" />
+        <field name="amount_python_compute">
+result = ((inputs.SALEURO and inputs.SALEURO.amount) + (inputs.SALASIA and
+inputs.SALASIA.amount)) * 0.01
+        </field>
+    </record>
+    <!-- Rule Inputs -->
+    <record id="hr_rule_input_sale_a" model="hr.rule.input">
+        <field name="code">SALEURO</field>
+        <field name="name">Sales to Europe</field>
+        <field name="input_id" ref="hr_salary_rule_sales_commission" />
+    </record>
+    <record id="hr_rule_input_sale_b" model="hr.rule.input">
+        <field name="code">SALASIA</field>
+        <field name="name">Sales to Asia</field>
+        <field name="input_id" ref="hr_salary_rule_sales_commission" />
+    </record>
+</odoo>
