+++ payroll_salary_rule_python/models/__init__.py
@@ -1,0 +1,1 @@
+from . import hr_salary_rule
