+++ payroll_salary_rule_python/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Henrik Norlin <henrik@appstogrow.co>
+* Those who have contributed to "payroll"
