+++ payroll_salary_rule_python/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module implements user-defined python code.
+It is moved here, away from "payroll".
+Therefore the licence is LGPL-3, the same as for "payroll".
