+++ payroll_salary_rule_python/readme/USAGE.rst
@@ -1,0 +1,1 @@
+See the explanation in `condition_python` and `amount_python_compute`.
