+++ payroll_salary_rule_python/views/hr_salary_rule_views.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="hr_salary_rule_view_form" model="ir.ui.view">
+        <field name="name">hr.salary.rule.form</field>
+        <field name="model">hr.salary.rule</field>
+        <field name="inherit_id" ref="payroll.hr_salary_rule_view_form" />
+        <field name="arch" type="xml">
+            <field name="condition_select" position="after">
+                <field
+                    name="condition_python"
+                    attrs="{'invisible':[('condition_select','!=','python')], 'required': [('condition_select','=','python')]}"
+                    colspan="4"
+                    widget="ace"
+                    options="{'mode': 'python'}"
+                    id="condition_python"
+                    nolabel="1"
+                />
+            </field>
+            <field name="amount_fix" position="after">
+                <field
+                    colspan="4"
+                    name="amount_python_compute"
+                    attrs="{'invisible':[('amount_select','!=','code')], 'required':[('amount_select','=','code')]}"
+                    widget="ace"
+                    options="{'mode': 'python'}"
+                    id="amount_python_compute"
+                    nolabel="1"
+                />
+            </field>
+        </field>
+    </record>
+</odoo>
