+++ setup/payroll_salary_rule_python/odoo/addons/payroll_salary_rule_python
@@ -1,0 +1,1 @@
+../../../../payroll_salary_rule_python
